backend/pyproject.toml
[tool.poetry]
name = "django-vue-docker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9.6"
Django = "^4.2.3"
gunicorn = "^20.1.0"
poetry-dotenv-plugin = "^0.1.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
django-cors-headers = "^4.1.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
